Set value to unknown when new_value is set
Since new_value is a write-only attribute, Terraform does not guarantee
that it won't change between plan and apply. Since our "value" argument
is _not_ write-only, we need to set it to an unknown value during planning
so that we can decide its final value based on whatever new_value is set
to during the apply phase.

This also includes some other general robustness improvements.